+++ yukkurisim/yukkurisim/KariAction.java
@@ -264,7 +264,7 @@ public class KariAction extends ItemGeneratorActionBase {
 	@Override
 	public int getIndexLevel(int status) {
 		// 基本コレでよいような気がするんだがなあ・・・
-		return (int)(status/UniqueTable.size());
+		return status/UniqueTable.size();
 	}
 
 	@Override
